pyproject.toml
[project]
name = "titan"
version = "0.1.0"
description = "TITAN — Testing Interfaces, Transactions And Networks"
requires-python = ">=3.9"
dependencies = [
    "playwright>=1.49",
    "pyyaml>=6.0",
    "pillow>=10.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.5"]
anthropic = ["anthropic>=0.40"]
openai = ["httpx>=0.27"]

[project.scripts]
titan = "cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B"]
ignore = [
    "E501",   # line length
    "E701",   # multiple statements on one line (concise conditionals in test code)
    "E741",   # ambiguous variable name (l, I, O — used in loops)
    "F841",   # unused variable (intermediate screenshots assigned for clarity)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S"]
